renamed "features" folder to "services", according to the term TypirServices
